refactor(cli): address review — pure BuildRegistry, single-gate manifest docs
Addresses the review comments on #673:

- BuildRegistry takes importMergeEnabled as a parameter instead of reading
  config.GetConfig() internally, so it stays a pure function of its args and is
  testable without touching global viper state. The flag is resolved at the 3
  call sites (project.registry, deps.Registry, ruledoc.Build).
- Move manifest-inline-conflict.docs.yaml under the import-manifest provider's
  docs FS so the rule and its doc fragment ride the same gate. Drops the fragile
  e.RuleID == "project/manifest-inline-conflict" string match in ruledoc.
- buildregistry_test drops the viper helper and passes the bool directly; the
  helper moves to manifest_wiring_test, whose New/Load path genuinely reads
  global config.


🔒 Scanned for secrets using gitleaks 8.28.0
