+++ cli/internal/app/dependencies.go
@@ -257,7 +257,7 @@ func (d *deps) CompositeProvider() provider.Provider {
 // sharing the same construction as project validation so the docs generator
 // observes an identical rule set.
 func (d *deps) Registry() (rules.Registry, error) {
-	return project.BuildRegistry(d.CompositeProvider(), importmanifest.New())
+	return project.BuildRegistry(d.CompositeProvider(), importmanifest.New(), config.GetConfig().ExperimentalFlags.ImportMerge)
 }
 
 // NewProject creates a project with composite provider.
